requirements.txt
asgiref
atomicwrites
attrs
Babel
certifi
charset-normalizer
colorama
colorlog
coreapi
coreschema
coverage
reportlab
Django
django-cors-headers
django-filter
django-sequences
django-simple-history
djangorestframework
drf-yasg2
drf-yasg
execnet
gunicorn
idna
inflection
iniconfig
itypes
Jinja2
MarkupSafe
packaging
pandas
pluggy
psycopg2-binary
py
PyJWT
pyparsing
pytest
pytest-django
pytest-forked
pytest-xdist
pytz
requests
requests-mock
sqlparse
tomli
uritemplate
urllib3
python-dateutil
num2words
xhtml2pdf
Numpy_financial
openpyxl
factory-boy
responses
django-extensions
pypdf
Pillow
PyMuPDF
pytest-cov
drf-writable-nested==0.7.0
setuptools
django-cors-headers

